+++ controllers/middlewares/expenses.validation.js
@@ -1,6 +1,6 @@
 const stringIsFloat = require("../utils/validation/stringIsFloat");
 const stringIsInteger = require("../utils/validation/stringIsInteger");
-const getMonth = require("../../utils/getMonth");
+const getMonth = require("../../src/utils/getMonth");
 const isValidType = require("../utils/validation/validType");
 const isValidTransactionType = require("../utils/validation/isValidTransactionType");
 const expenseModel = require("../../models/expense.model");
