+++ controllers/setup.controller.js
@@ -1,7 +1,7 @@
 const setUpModel = require('../models/setup.model');
 const {set} = require("express/lib/application");
 const {parse} = require("nodemon/lib/cli");
-const expensesTypes = require('../utils/types');
+const expensesTypes = require('../src/utils/types');
 /**
  * Retrieves the setup for a given year
  * @param req
