+++ controllers/utils/validation/isValidTransactionType.js
@@ -1,4 +1,4 @@
-const transactionTypes = require('../../../utils/transactionTypes');
+const transactionTypes = require('../../../src/utils/transactionTypes');
 
 function isValidTransactionType(transaction) {
     return transaction === transactionTypes.EXPENSE || transaction === transactionTypes.INCOME;
