+++ controllers/utils/validation/validType.js
@@ -1,4 +1,4 @@
-const expensesTypes = require('../../../utils/types');
+const expensesTypes = require('../../../src/utils/types');
 
 /**
  * Validates that the type is one of the types in types.js
