+++ src/utils/queryExpensesOfAType.js
